backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
mysqlclient>=2.2
python-dotenv>=1.0
django-cors-headers>=4.3
